# service/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Инспектор ИИ",
    "version": "0.1.0",
    "description": "Сверка проектной, рабочей и исполнительной документации. Асинхронная pull-модель: загрузка возвращает process_id, клиент опрашивает статус и забирает протокол."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "FindingCounts": {
        "type": "object",
        "description": "Счётчики для пяти таблиц протокола ТЗ: кандидаты и решения по ним; no_violation — проверено, расхождения нет; not_comparable — сравнение невозможно; split — составные записи, разделённые инспектором; hypotheses — записи свободного поиска, не переведённые в кандидаты: они не входят в candidates и pending.",
        "required": [
          "total",
          "candidates",
          "pending",
          "hypotheses",
          "confirmed",
          "rejected",
          "clarification",
          "split",
          "no_violation",
          "not_comparable"
        ],
        "properties": {
          "total": {
            "type": "integer"
          },
          "candidates": {
            "type": "integer"
          },
          "pending": {
            "type": "integer"
          },
          "hypotheses": {
            "type": "integer"
          },
          "confirmed": {
            "type": "integer"
          },
          "rejected": {
            "type": "integer"
          },
          "clarification": {
            "type": "integer"
          },
          "split": {
            "type": "integer"
          },
          "no_violation": {
            "type": "integer"
          },
          "not_comparable": {
            "type": "integer"
          },
          "revisit": {
            "type": "integer",
            "description": "Записи с решением инспектора, у которых после дозагрузки изменились значения: решение принято по прежним значениям и его надо пересмотреть. Новое решение отметку снимает."
          },
          "by_section": {
            "type": "object",
            "description": "По разделам Матрицы (ключ — начало кода параметра: PZ, SPZU, AR, KR, IOS1…IOS5, POS, POD, OOS, PPM, ODI, ZU, SM): сколько записей ждут решения и сколько подтверждено. Разделы без таких записей не перечисляются.",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "candidates": {
                  "type": "integer"
                },
                "confirmed": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
...
      "Finding": {
        "type": "object",
        "required": [
          "id",
          "finding_id",
          "verification_status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "finding_id": {
            "type": "string"
          },
          "protocol_version": {
            "type": "integer"
          },
          "parameter_code": {
            "type": "string",
            "nullable": true
          },
          "parameter_id": {
            "type": "integer",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "violation_label": {
            "type": "string",
            "nullable": true
          },
          "protocol_status": {
            "type": "string",
            "nullable": true
          },
          "criticality": {
            "type": "string",
            "nullable": true
          },
          "pd_value": {
            "type": "string",
            "nullable": true
          },
          "rd_value": {
            "type": "string",
            "nullable": true
          },
          "id_value": {
            "type": "string",
            "nullable": true,
            "description": "Значение исполнительной документации (#42)"
          },
          "id_check": {
            "type": "string",
            "nullable": true,
            "description": "Итог сверки с исполнительной документацией (#42)"
          },
          "sources": {
            "type": "object",
            "description": "Чем прочитано значение стороны и к чему привязано (#41): text_source RECOGNIZED — распознано машиной (Tesseract или модель); binding CLAUSE — из указаний, SHEET — с листа чертежа, PAGE — со страницы, PATH — по пути файла",
            "properties": {
              "pd": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "text_source": {
                    "type": "string",
                    "nullable": true
                  },
                  "binding": {
                    "type": "string",
                    "nullable": true
                  }
                }
              },
              "rd": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "text_source": {
                    "type": "string",
                    "nullable": true
                  },
                  "binding": {
                    "type": "string",
                    "nullable": true
                  }
                }
              },
              "id": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "text_source": {
                    "type": "string",
                    "nullable": true
                  },
                  "binding": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          },
          "needs_expert": {
            "type": "boolean",
            "description": "Система не решает сама (#41): распознанное значение, расхождение со слоем, замена марки — решает инспектор"
          },
          "value_read_by": {
            "type": "object",
            "description": "RECOGNIZED у стороны, значение которой прочитано машиной, а не текстовым слоем (#52)",
            "properties": {
              "pd": {
                "type": "string",
                "nullable": true
              },
              "rd": {
                "type": "string",
                "nullable": true
              },
              "id": {
                "type": "string",
                "nullable": true
              }
            }
          },
          "finding_status": {
            "type": "string",
            "nullable": true,
            "description": "Статус по легенде организатора к Матрице с учётом решения инспектора (#80): CANDIDATE, CONFIRMED_VIOLATION, NEGATIVE_VERIFIED, CLARIFICATION_REQUIRED, MISSING_EVIDENCE, NOT_COMPARABLE, SUSPICION"
          },
          "specialist": {
            "type": "object",
            "nullable": true,
            "description": "Решение специалиста (#129, Р-86, Р-108): violation — подтверждено нарушением, need_info — специалисту не хватило данных, no_violation — не нарушение, hypothesis — вид показывать гипотезой, drop — специалист решил вид не показывать (по решению пользователя такие записи видны в конце таблицы гипотез; как исполнено — в note). scope — к чему решение: kind — к виду гипотезы, record — к этой записи с теми значениями, которые видел специалист",
            "required": [
              "verdict"
            ],
            "properties": {
              "verdict": {
                "type": "string",
                "enum": [
                  "violation",
                  "need_info",
                  "no_violation",
                  "hypothesis",
                  "drop"
                ]
              },
              "scope": {
                "type": "string",
                "enum": [
                  "kind",
                  "record"
                ]
              },
              "reviewed": {
                "type": "string",
                "nullable": true
              },
              "note": {
                "type": "string",
                "nullable": true
              },
              "cases": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "confidence": {
            "type": "object",
            "nullable": true,
            "description": "Уверенность в выводе (#80): уровень — для отбора, и признаки, из которых он сложен. Процента нет: уровень выводится из признаков записи и показывается вместе с ними. HIGH — признаков сомнения нет, MEDIUM — вывод опирается на то, что стоит проверить глазами (прочитано машиной, взято с листа, соперничающее значение), LOW — система не решает сама. null — сравнения не было",
            "required": [
              "level",
              "down",
              "up"
            ],
            "properties": {
              "level": {
                "type": "string",
                "enum": [
                  "HIGH",
                  "MEDIUM",
                  "LOW"
                ]
              },
              "down": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "что снижает уверенность"
              },
              "up": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "на что вывод опирается"
              }
            }
          },
          "verification_status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED_VIOLATION",
              "NEGATIVE_VERIFIED",
              "CLARIFICATION_REQUIRED",
              "NOT_REQUIRED",
              "SPLIT"
            ]
          },
          "reason_code": {
            "type": "string",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "decided_by": {
            "type": "string",
            "nullable": true
          },
          "decided_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "locations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "location_type": {
            "type": "string",
            "nullable": true
          },
          "comparison_result": {
            "type": "string",
            "nullable": true
          },
          "matrix_scope": {
            "type": "string",
            "nullable": true
          },
          "provisional": {
            "type": "boolean",
            "description": "Находка черновика правила (#95): параметр без правила прода проверен черновиком, который ждёт решения специалиста. Как гипотеза свободного поиска — вне итогов, пока инспектор не взял её в кандидаты"
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "rule_basis": {
            "type": "string",
            "nullable": true
          },
          "origin": {
            "type": "string",
            "nullable": true
          },
          "carried_from_version": {
            "type": "integer",
            "nullable": true,
            "description": "Запись перенесена из этой версии протокола (#60): дозагруженные документы её не касались, значения и решение инспектора относятся к тому разбору. null — запись пересчитана сейчас"
          },
          "parent_finding_id": {
            "type": "string",
            "nullable": true
          },
          "completeness_status": {
            "type": "string",
            "nullable": true,
            "description": "Комплектность записи по ТЗ 9.2: COMPLETE, MISSING_EVIDENCE, NOT_COMPARABLE"
          },
          "completeness_reason": {
            "type": "string",
            "nullable": true
          },
          "revisions": {
            "type": "object",
            "description": "Прежние редакции документов, в которых значение другое: сравнение идёт по последней редакции, устаревшие показываются для проверки (ТЗ 9.1)",
            "properties": {
              "pd": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              },
              "rd": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "changed_after_decision": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true,
            "description": "Значения до дозагрузки, если они изменились после решения инспектора: решение надо пересмотреть"
          },
          "reopened": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true,
            "description": "Прежнее решение инспектора, которое пересборка вернула ему в кандидаты: после дозагрузки автоматика расхождения больше не видит. decision, by, at, reason_code, comment, version — в какой версии протокола вернула, before — значения ПД и РД на момент решения. Новое решение отметку снимает."
          },
          "promoted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Когда инспектор взял гипотезу свободного поиска в кандидаты (ТЗ 9.2); до этого решения по ней не принимаются"
          },
          "promoted_by": {
            "type": "string",
            "nullable": true
          },
          "disputed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Когда инспектор вернул в кандидаты запись, проверенную автоматически без нарушения (#74): она снова ждёт решения и переживает пересборку протокола"
          },
          "disputed_by": {
            "type": "string",
            "nullable": true
          },
          "evidence": {
            "type": "array",
            "description": "Доказательства. Кроме полей сдачи: document_code и document_sheet_number, если лист определён; page_count; preview.aspect — пропорции страницы; highlights — прямоугольники места во внутренней системе координат (доли видимой страницы, Y сверху) и highlight_source — чем место найдено (VALUE, ROWS, LOCATION, FINDING — место указано самой находкой, REVISION — места, где лист изменился между редакциями, #81). Пустой highlights при localization PAGE_LEVEL или DOCUMENT_LEVEL означает, что доказательство относится к странице или документу целиком и обводить нечего; при BBOX — что место на странице не найдено.",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/findings/{id}/decision": {
      "post": {
        "summary": "Решение инспектора по кандидату",
        "tags": [
          "findings"
        ],
        "description": "ТЗ, раздел 9.3. CONFIRM — нарушение подтверждено; REJECT — отклонено, reason_code и комментарий обязательны; CLARIFY — требуется уточнение; RESET — вернуть решённую запись в кандидаты до финализации; PROMOTE — взять гипотезу свободного поиска в кандидаты; DISPUTE — вернуть в кандидаты запись, которую автоматика закрыла сама как «расхождения нет» (#74). Комментарий инспектора можно оставить при любом решении, он попадает в журнал аудита и в протокол вместе с пользователем и временем. В статусе COMPLETED (верификация завершена) принимается только RESET: по таблице статусов ТЗ 9.1 решения там не принимаются.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "action"
                ],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "CONFIRM",
                      "REJECT",
                      "CLARIFY",
                      "RESET",
                      "PROMOTE",
                      "DISPUTE"
                    ]
                  },
                  "reason_code": {
                    "type": "string",
                    "enum": [
                      "WRONG_REVISION",
                      "APPROVED_CHANGE",
                      "OCR_ERROR",
                      "LINKING_ERROR",
                      "NOT_APPLICABLE"
                    ]
                  },
                  "comment": {
                    "type": "string",
                    "maxLength": 4000
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "finding",
                    "process_status",
                    "counts"
                  ],
                  "properties": {
                    "finding": {
                      "$ref": "#/components/schemas/Finding"
                    },
                    "process_status": {
                      "type": "string"
                    },
                    "counts": {
                      "$ref": "#/components/schemas/FindingCounts"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/findings/{id}/split": {
      "post": {
        "summary": "Разделение составного кандидата на атомарные",
        "tags": [
          "findings"
        ],
        "description": "ТЗ и задача #33: декомпозиция составного нарушения (например, несколько помещений в одном кандидате) на отдельные атомарные записи с сохранением доказательной базы.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "locations"
                ],
                "properties": {
                  "locations": {
                    "type": "array",
                    "minItems": 2,
                    "items": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "comment": {
                    "type": "string",
                    "maxLength": 1000
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "original_finding_id",
                    "created_findings",
                    "counts",
                    "process_status"
                  ],
                  "properties": {
                    "original_finding_id": {
                      "type": "string"
                    },
                    "created_findings": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Finding"
                      }
                    },
                    "counts": {
                      "$ref": "#/components/schemas/FindingCounts"
                    },
                    "process_status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}